Disintegrate whole scanner into separate classes
